requirements.txt
pandas
llama-index-core
llama-index-retrievers-bm25
python-dotenv
google-generativeai